# doc/api.yaml
openapi: 3.0.0
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Freddie Mercury'
        grammy:
          type: boolean
          example: false
      required:
        - name
        - grammy
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Innuendo'
        year:
          type: integer
          minimum: 1
          example: 1991
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - name
        - year
        - artistId
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: The Show Must Go On
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: integer
          minimum: 1
          description: In seconds
          example: 262
      required:
        - name
        - duration
        - artistId
        - albumId
...
paths:
...
  /track/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Track
      summary: Gets single track by id
      description: Gets single track by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        400:
          description: Bad request (id (trackId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Track was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    put:
      tags:
        - Track
      summary: Updates track information
      description: Updates library track information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                  example: Bohemian Rhapsody
                artistId:
                  type: string
                  format: uuid
                  nullable: true
                duration:
                  type: integer
                  description: In seconds
                  example: 262
                  minimum: 1
                albumId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        200:
          description: The track has been updated.
          content:
            schema:
              $ref: '#/components/schemas/Track'
        400:
          description: Bad request (id (trackId) is invalid (not uuid) or body is empty or the data type of the fields does not match the required or track name is empty string / duration is equal 0)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Track was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    delete:
      tags:
        - Track
      summary: Deletes track
      description: Deletes track from library
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request (id (trackId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Track was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
...
  /album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Album
      summary: Gets single album by id
      description: Gets single album by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request(id (albumId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    put:
      tags:
        - Album
      summary: Updates album information
      description: Updates library album information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                year:
                  type: integer
                artistId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        200:
          description: The album has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request (id (albumId) is invalid (not uuid) or body is empty or the data type of the fields does not match the required or album name is empty string / duration is equal 0)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    delete:
      tags:
        - Album
      summary: Deletes album
      description: Deletes album from library
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request (id (albumId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
...
  /artist/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Artist
      summary: Gets single artist by id
      description: Gets single artist by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        400:
          description: Bad request (id (artistId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Artist was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    put:
      tags:
        - Artist
      summary: Updates artist information
      description: Updates artist information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Freddie Mercury'
                grammy:
                  type: boolean
                  example: true
      responses:
        200:
          description: The artist has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        400:
          description: Bad request (id (artistId) is invalid (not uuid) or body is empty or the data type of the fields does not match the required or artist name is empty string)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Artist was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    delete:
      tags:
        - Artist
      summary: Deletes artist
      description: Deletes artist from library
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request (id (artistId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Artist was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
...
  /favs/track/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Adds track to the favorites
      description: Adds track to the favorites
      responses:
        201:
          description: Added successfully
        400:
          description: Bad request (id (track) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        422:
          description: Track with id doesn't exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    delete:
      tags:
        - Favorites
      summary: Deletes track from favorites
      description: Deletes track from favorites
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request (id (trackId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Track was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
  /favs/album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Adds album to the favorites
      description: Adds album to the favorites
      responses:
        201:
          description: Added successfully
        400:
          description: Bad request (id (albumId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        422:
          description: Album with id doesn't exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    delete:
      tags:
        - Favorites
      summary: Deletes album from favorites
      description: Deletes album from favorites
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request (id (albumId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
  /favs/artist/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Adds artist to the favorites
      description: Adds artist to the favorites
      responses:
        201:
          description: Added successfully
        400:
          description: Bad request (id (artistId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        422:
          description: Artist with id doesn't exist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
    delete:
      tags:
        - Favorites
      summary: Deletes artist from favorites
      description: Deletes artist from favorites
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request (id (artistId) is invalid (not uuid))
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'
        404:
          description: Artist was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Error message content'